+++ generator/src/main/kotlin/com/kylemayes/generator/generate/file/Extensions.kt
@@ -7,7 +7,21 @@ import com.kylemayes.generator.generate.support.generateManualUrl
 import com.kylemayes.generator.generate.support.getExtensionGroups
 import com.kylemayes.generator.registry.Extension
 import com.kylemayes.generator.registry.Registry
+import com.kylemayes.generator.registry.intern
 import com.kylemayes.generator.support.toPascalCase
+
+/** The warning and `cfg` applied to the documentation for provisional extensions. */
+private val PROVISIONAL: String =
+    """
+///
+/// ## WARNING
+///
+/// This is a
+/// [provisional extension](https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/provisional-headers.html).
+/// Provisional extensions are not guaranteed to be backwards compatible and are
+/// not intended to be used in production applications.
+#[cfg(feature = "provisional")]
+   """.trimEnd()
 
 /** Generates Rust modules and constants for Vulkan extensions. */
 fun Registry.generateExtensions() =
@@ -42,9 +56,10 @@ pub const fn to_extension_name(bytes: &[u8]) -> ExtensionName {
 
 /** Generates a Rust constant for a Vulkan extension. */
 private fun Registry.generateExtension(extension: Extension): String {
+    val provisional = if (extension.provisional) { PROVISIONAL } else { "" }
     val deprecation = generateDeprecation(extension)?.let { "\n$it" } ?: ""
     return """
-/// <${generateManualUrl(extension)}>$deprecation
+/// <${generateManualUrl(extension)}>$provisional$deprecation
 pub const ${extension.name}_EXTENSION: ExtensionName =
     to_extension_name(b"${extension.name.original}");
     """.trim()
@@ -84,22 +99,66 @@ impl ConvertCStr for ExtensionName {
     }
 }
 
+/// A collection of metadata for a Vulkan extension.
+#[derive(Copy, Clone, Debug)]
+pub struct ExtensionMetadata {
+    /// The name of the extension.
+    pub name: ExtensionName,
+    /// The number assigned to the extension.
+    pub number: i32,
+
+    /// The type of the extension (`device` or `instance`).
+    pub type_: &'static str,
+
+    /// The author of the extension.
+    pub author: &'static str,
+    /// The primary contact(s) for the extension.
+    pub contact: &'static str,
+
+    /// The platform the extension can be used on (e.g., `wayland` or `win32`).
+    pub platform: Option<&'static str>,
+
+    /// The extensions required by the extension.
+    pub requires: Option<&'static [ExtensionName]>,
+    /// The Vulkan version required by the extension (e.g., `1.1`).
+    pub requires_core: Option<&'static str>,
+
+    /// The Vulkan extension or version that deprecated the extension (e.g., `VK_VERSION_1_1`).
+    pub deprecated_by: Option<&'static str>,
+    /// The Vulkan extension or version that obsoleted the extension (e.g., `VK_VERSION_1_1`).
+    pub obsoleted_by: Option<&'static str>,
+    /// The Vulkan version the extension was promoted to core in (e.g., `VK_VERSION_1_1`).
+    pub promoted_to: Option<&'static str>,
+}
+
 ${getExtensionGroups().values.flatten().sortedBy { it.name }.joinToString("") { generateExtensionTrait(it) }}
     """
-
 /** Generates a Rust trait and implementation for a Vulkan extension. */
 private fun Registry.generateExtensionTrait(extension: Extension): String {
+    val provisional = if (extension.provisional) { PROVISIONAL } else { "" }
     val deprecation = generateDeprecation(extension)?.let { "\n$it" } ?: ""
+
     val name = "${extension.name.value.toPascalCase()}Extension"
     val type = extension.type!!.capitalize()
+
     val commands = extension.require.commands.mapNotNull { commands[it] }.sortedBy { it.name }
+
+    val implAttributes = listOf(
+        if (extension.provisional) { "#[cfg(feature = \"provisional\")]" } else { "" },
+        if (deprecation.isNotEmpty()) { "#[allow(deprecated)]" } else { "" }
+    ).filter { it.isNotBlank() }.joinToString("\n")
+
     return """
-/// <${generateManualUrl(extension)}>$deprecation
+/// <${generateManualUrl(extension)}>$provisional$deprecation
 pub trait $name: ${type}V1_0 {
+    /// The metadata for this extension.
+    #[allow(deprecated)]
+    const METADATA: ExtensionMetadata = ${generateMetadata(extension)};
+
     ${commands.joinToString("") { generateCommandWrapper(it) }}
 }
 
-${if (deprecation.isNotEmpty()) { "#[allow(deprecated)]" } else { "" }}
+$implAttributes
 impl $name for crate::$type { }
     """
 }
@@ -111,4 +170,33 @@ private fun generateDeprecation(extension: Extension) = extension.deprecatedby?.
     } else {
         "#[deprecated]"
     }
+}
+
+/** Generates a Rust struct for the metadata for a Vulkan extension. */
+private fun Registry.generateMetadata(extension: Extension): String {
+    val requires = if (extension.requires != null) {
+        val names = extension.requires
+            .split(",")
+            .map { extensions[it.intern()]!!.name }
+            .joinToString(", ") { "super::${it}_EXTENSION" }
+        "Some(&[$names])"
+    } else {
+        "None"
+    }
+
+    return """
+ExtensionMetadata {
+    name: super::${extension.name}_EXTENSION,
+    number: ${extension.number},
+    type_: "${extension.type}",
+    author: "${extension.author}",
+    contact: "${extension.contact}",
+    platform: ${extension.platform?.let { "Some(\"$it\")" } ?: "None"},
+    requires: $requires,
+    requires_core: ${extension.requiresCore?.let { "Some(\"$it\")" } ?: "None"},
+    deprecated_by: ${extension.deprecatedby?.let { "Some(\"$it\")" } ?: "None"},
+    obsoleted_by: ${extension.obsoletedby?.let { "Some(\"$it\")" } ?: "None"},
+    promoted_to: ${extension.promotedto?.let { "Some(\"$it\")" } ?: "None"},
+}
+    """
 }