+++ generator/src/main/kotlin/com/kylemayes/generator/generate/support/Wrapper.kt
@@ -81,13 +81,25 @@ fun Registry.generateCommandWrapper(command: Command): String {
 
                 if (outputWithLength || outputWithoutLength) {
                     // Output slice parameter.
-                    val length = if (outputWithLength) { current.name.value } else { "${slices[0].name}.len()" }
-                    resultTypes.add("Vec<${pointee.generate()}>")
+
+                    val length = if (outputWithLength) {
+                        current.name.value
+                    } else {
+                        "${slices[0].name}.len()"
+                    }
+
+                    if (pointee.getIdentifier()?.value == "void") {
+                        resultTypes.add("Vec<u8>")
+                        addArgument("${slice.name}.as_mut_ptr() as *mut c_void", setup = "ptr::null_mut()")
+                    } else {
+                        resultTypes.add("Vec<${pointee.generate()}>")
+                        addArgument("${slice.name}.as_mut_ptr()", setup = "ptr::null_mut()")
+                    }
+
                     resultExprs.add(slice.name.value)
                     preActualStmts.add("let mut ${slice.name} = Vec::with_capacity($length as usize);")
                     postActualStmts.add("debug_assert!(${slice.name}.capacity() == $length as usize);")
                     postActualStmts.add("unsafe { ${slice.name}.set_len($length as usize) };")
-                    addArgument("${slice.name}.as_mut_ptr()", setup = "ptr::null_mut()")
                 } else {
                     // Input slice parameter.